requirements.txt
requests == 2.32.3
pytest == 8.3.4
python-dotenv == 1.0.1
behave == 1.2.6